dataframe data types -- changed age to a float for the titanic data

+++ tests/test_dataframe.py
@@ -155,7 +155,7 @@ data_types = {
     "Pclass": int,
     "Name": str,
     "Sex": str,
-    "Age": int,
+    "Age": float,
     "SibSp": int,
     "Parch": int,
     "Ticket": str,
